SendEmail, update routines and Encrypt password

+++ src/main/java/com/example/proyectocomponentes/models/Trainer.java
@@ -19,10 +19,11 @@ public class Trainer {
     private String lastName;
     @Column(nullable = false)
     private String email;
+
     @Column(nullable = false)
     private String password;
     @Column(nullable = false)
-    private Integer phone;
+    private Long phone;
 
     @JsonIgnore
     @OneToMany(mappedBy = "trainer", cascade = CascadeType.ALL)
